change  /image-sets/ end points
